refactor: make SignUpStage 1 and 2 components
